add some TODO comments and move code into appropriate preprocessor
